Extract mouse input press action execution
Move click and long-press sequencing, runtime coordinate conversion,
and button mapping behind one stateless executor so the use case keeps
only tool orchestration. Preserve the async source guard for the move.
